Update UI for subdocs uploading
